+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### New
+
+- cmd/lint-changelog: handle french date format.
+
 ## [0.5.0] - 2025-10-28
 
 ### Changed